move utils to utils package + move audio functions to new audio_utils module
